Add a new javascript file
